Ref #774 : Fix tests
